Organised the weather page into a separate file, moving the placeholder searchLocation function with it.
