+++ lib/home_screen.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'widgets/widgets.dart';
+import 'weather_screen.dart';
 
 class HomePage extends StatefulWidget {
   const HomePage({required this.username, super.key});
@@ -31,10 +32,6 @@ class _HomePage extends State<HomePage> {
 
   void onCancelButtonPressed() {
     print('Cancel button pressed.');
-  }
-
-  void searchLocation(String location) {
-    print('Searching for $location');
   }
 
   @override
@@ -140,81 +137,7 @@ class _HomePage extends State<HomePage> {
                 ),
             ) : selectedIndex == 1
                 // Weather Page -----------------------------------------------------------------
-                ? Container(
-                  color: const Color(0xFFACD4AE),
-                  child: Column(
-                    children: [
-                      Padding(
-                        padding: const EdgeInsets.all(20.0),
-                        child: TextField( // TODO: Implement a text input widget to reuse
-                          decoration: InputDecoration(
-                            hintText: 'Search Location',
-                            border: OutlineInputBorder(),
-                          ),
-                          
-                          onChanged: (value) {
-                            searchLocation(value);
-                          },
-                        ),
-                      ),
-                      Padding(
-                        padding: const EdgeInsets.only(left: 20.0),
-                        child: Row(
-                          children: [
-                            Text('London', style: Theme.of(context).textTheme.headlineLarge),
-                          ],
-                        ),
-                      ),
-                      Padding(
-                        padding: const EdgeInsets.only(left: 20.0),
-                        child: Row(
-                          children: [
-                            Text('5°C', style: Theme.of(context).textTheme.titleLarge!.copyWith(fontSize: 80.0)),
-                            SizedBox(width: 10),
-                            Text('Cloudy', style: Theme.of(context).textTheme.bodyMedium!.copyWith(fontSize: 32.0)),
-                          ],
-                        ),
-                      ),
-                      Padding(
-                        padding: const EdgeInsets.only(left: 20.0),
-                        child: Row(
-                          children: [
-                            Text('Feb 11, 2025', style: Theme.of(context).textTheme.bodyLarge!.copyWith(fontSize: 20.0)),
-                          ],
-                        ),
-                      ),
-                      Padding(
-                        padding: const EdgeInsets.all(20.0),
-                        child: Row(
-                          children: [
-                            Expanded(
-                              child: Container(
-                                height: 104,
-                                decoration: BoxDecoration(
-                                  color: Colors.white,
-                                  borderRadius: BorderRadius.circular(10),
-                                ),
-                                child: ListView(
-                                  scrollDirection: Axis.horizontal,
-                                  children: [
-                                    HourlyTemperature(time: '7:00', temperature: '5°C'),
-                                    HourlyTemperature(time: '8:00', temperature: '5°C'),
-                                    HourlyTemperature(time: '9:00', temperature: '5°C'),
-                                    HourlyTemperature(time: '10:00', temperature: '5°C'),
-                                    HourlyTemperature(time: '11:00', temperature: '6°C'),
-                                    HourlyTemperature(time: '12:00', temperature: '7°C'),
-                                    HourlyTemperature(time: '13:00', temperature: '8°C'),
-                                    HourlyTemperature(time: '14:00', temperature: '7°C'),
-                                  ],
-                                )
-                              ),
-                            )
-                          ],
-                        ),
-                      )
-                    ],
-                  ),
-                ) 
+                ? WeatherPage()
                 // Profile Page -----------------------------------------------------------------
                 : Container(
                   color: const Color(0xFFACD4AE),
